+++ app/src/main/java/io/familymoments/app/feature/home/screen/HomeScreen.kt
@@ -45,7 +45,7 @@ import io.familymoments.app.feature.home.viewmodel.HomeViewModel
 fun HomeScreen(
     modifier: Modifier,
     viewModel: HomeViewModel,
-    navigateToPostDetail: () -> Unit
+    navigateToPostDetail: (Int) -> Unit
 ) {
     val homeUiState = viewModel.homeUiState.collectAsStateWithLifecycle()
     val posts = homeUiState.value.posts
